Menu Rewrite complete
Main functionality seems to be good and it's ready for general use.

May still be some bugs, and some events might not be right, I still need to verify many of them, I mostly tested with the Applications Event Class because i had a bunch of those in my event viewer.  The AD ones aren't set right, but I'll get to that soon.